# artifacts/openapi.yaml
openapi: 3.0.3
info:
  title: URL Shortener API
  version: 1.0.0
  description: |
    Signed-in customers turn long URLs into short links; anyone may click one and
    is redirected; every click is counted exactly and reported back to the link's
    owner.

    This document is the frozen contract. The implementation and the tests are
    written in parallel, in isolation, by authors who never see each other's
    work, so anything not fixed here is something they can disagree about. What
    follows is therefore more prescriptive than an API description usually is.

    ## Two surfaces

    | Surface | Path | Authentication |
    |---|---|---|
    | Management API | `/api/v1/**`, JSON | `Authorization: Bearer <token>`, except sign-in |
    | Click path | `/{code}` at the root | none, ever (AC14) |

    The redirect lives at the root and not under a prefix because that is what
    makes a short link short. Uniqueness of a code is scoped to
    `(domain, code)` from day one, with a single configured domain in this
    build, so customer-owned domains arrive later as data rather than as a
    migration.

    ## The single 404

    Unknown code, expired link, deleted link, blocked link, and a link that
    exists but belongs to another customer all answer

    ```
    404 Not Found
    Content-Type: application/json
    {"error":"not_found","message":"Not found"}
    ```

    byte for byte, on both surfaces. A malformed code answers the same way. There
...

    ## Error bodies

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/links/{code}/abuse-reports:
    parameters:
      - $ref: '#/components/parameters/CodePath'
    post:
      tags: [abuse]
      operationId: reportAbuse
      summary: Report a link as abusive
      description: |
        A report blocks the link immediately (AC9/AC21) - there is no moderation
        console in this build, so there is nobody for a queued report to wait for.
        The reporter's customer id, the timestamp and the optional reason are
        recorded for the review process that comes later.

        Reporting requires a signed-in session and is rate-limited per reporter.
        That is the whole defence against this endpoint being a cheap way to kill
        a competitor's links, and it is why an open, unauthenticated report
        endpoint with automatic takedown is deliberately not built. A reviewer
        should weigh that trade: this design chooses "a signed-in customer can
        take down any link they can name, subject to a rate limit and an audit
        row" over "abusive links stay up until a human is hired".

        The response is 202 for any syntactically acceptable code, whether or not
        a link exists and whoever owns it. Answering 404 for an unknown code would
        turn this endpoint into the existence oracle that every other endpoint is
        careful not to be.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AbuseReportRequest'
            example:
              reason: Phishing page imitating a bank sign-in
      responses:
        '202':
          description: |
            Report accepted. No body, and the same response whether or not the
            code resolves.
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        The token from `POST /api/v1/sessions`, verified locally against a public
        key. No request causes a call back to a login system (AC18).
  parameters:
    CodePath:
      name: code
      in: path
      required: true
      description: |
        The short code. Case-sensitive. No syntactic validation is applied on
        either surface: a code that could not possibly exist answers 404 like any
        other code that does not resolve, because a 400 for a malformed code would
        tell an enumerator which shapes are worth trying.
      schema:
        type: string
        maxLength: 64
      example: 7Qk2mZa9Xr4Lb0Nc8Tv1Ps
...
  responses:
    InvalidRequest:
      description: The request body or parameters are not valid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            error: invalid_request
            message: The request is not valid.
    Unauthorized:
      description: |
        No session token, or one that is malformed, expired or not verifiable.
        The reason is never disclosed.
      headers:
        WWW-Authenticate:
          description: '`Bearer`'
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            error: unauthorized
            message: Authentication required.
...
    RateLimited:
      description: A token bucket for this caller is empty.
      headers:
        Retry-After:
          description: Whole seconds, never 0.
          required: true
          schema:
            type: integer
            minimum: 1
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            error: rate_limited
            message: Too many requests.
  schemas:
    ApiError:
      type: object
      description: The shape of every non-2xx body in this API.
      required: [error, message]
      additionalProperties: false
      properties:
        error:
          type: string
          description: Machine-readable code from the closed catalogue.
          enum:
            - invalid_request
            - invalid_credentials
            - unauthorized
            - not_found
            - alias_unavailable
            - link_not_modifiable
            - url_rejected
            - rate_limited
            - service_unavailable
        message:
          type: string
          description: Fixed per code. Never echoes request content.
        fields:
          type: object
          description: |
            Per-field detail, present only on `invalid_request` and omitted
            entirely otherwise - which is what keeps the not-found body identical
            across every endpoint.
          additionalProperties:
            type: string
...
    AbuseReportRequest:
      type: object
      additionalProperties: false
      properties:
        reason:
          type: string
          maxLength: 500
...